fix/non-usd-charts (#278)
* and hopefully don't break anything else
