+++ apps/dapp/src/context/app-providers.tsx
@@ -3,14 +3,17 @@ import { EvmProvider } from "./evm-provider";
 import { ReactQueryProvider } from "./react-query-provider";
 import { HashRouter as Router } from "react-router-dom";
 import { MarketProvider } from "./market-context";
+import { TokenProvider } from "./token-context";
 
 export const Providers: FC<{ children: ReactNode }> = ({ children }) => {
   return (
     <ReactQueryProvider>
       <EvmProvider>
-        <MarketProvider>
-          <Router>{children}</Router>
-        </MarketProvider>
+        <TokenProvider>
+          <MarketProvider>
+            <Router>{children}</Router>
+          </MarketProvider>
+        </TokenProvider>
       </EvmProvider>
     </ReactQueryProvider>
   );
